Reorganize rdkitgridfeaturizer module
